% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_polygons.R
\docType{data}
\name{aus_polygons}
\alias{aus_polygons}
\title{Australian soil texture polygons (aus_polygons)}
\format{
dataframe-tibble
}
\source{
NCST 'Yellow Book' 4e, 2024
}
\usage{
aus_polygons
}
\description{
Data describing polygons for each Australian soil textural
  class. Note that the silt/sand divide for this system is set at 0.02 mm.
}
\keyword{datasets}
