% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_polygons.R
\docType{data}
\name{nz_polygons}
\alias{nz_polygons}
\title{New Zealand soil texture polygons (nz_polygons)}
\format{
dataframe-tibble
}
\source{
Soil description handbook Milne, J. D. G.; Clayden, B.; Singleton, P. L.; Wilson, A. D. 1995
}
\usage{
nz_polygons
}
\description{
Data describing polygons for each soil textural class
}
\keyword{datasets}
