% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzg_polygons.R
\docType{data}
\name{nzg_polygons}
\alias{nzg_polygons}
\title{New Zealand soil texture group polygons (nzg_polygons)}
\format{
dataframe-tibble
}
\source{
Soil description handbook Milne, J. D. G.; Clayden, B.; Singleton, P. L.; Wilson, A. D. 1995
}
\usage{
nzg_polygons
}
\description{
Data describing polygons for each soil texture group following Milne et al 1995
}
\keyword{datasets}
