% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssew_polygons.R
\docType{data}
\name{ssew_polygons}
\alias{ssew_polygons}
\title{Soil Survey of England and Wales soil texture polygons (ssew_polygons)}
\format{
dataframe-tibble
}
\source{
Natural England TIN037
}
\usage{
ssew_polygons
}
\description{
Data describing polygons for each soil textural class following Natural England TIN037
}
\keyword{datasets}
