% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss_polygons.R
\docType{data}
\name{swiss_polygons}
\alias{swiss_polygons}
\title{Austrian soil texture polygons (usda_polygons)}
\format{
dataframe-tibble
}
\source{
ÖSTERREICHISCHE BODENKARTIERUNG, 1967:
}
\usage{
swiss_polygons
}
\description{
Data describing polygons for each soil textural class
}
\keyword{datasets}
