% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usda_polygons.R
\docType{data}
\name{usda_polygons}
\alias{usda_polygons}
\title{USDA soil texture polygons (usda_polygons)}
\format{
dataframe-tibble
}
\source{
USDA and ggtern data(USDA)
}
\usage{
usda_polygons
}
\description{
Data describing polygons for each soil textural class
}
\keyword{datasets}
